yasp: Protocol and unittest updates.
Protocol now uses 0x05 as the frame start for future autobaud support.
Two bytes of message length and command length are also implemented.
Cleaned up some undefined constants and cleaned unit tests.

+++ yasp.h
@@ -10,11 +10,11 @@
 typedef void(*command_callback)(uint8_t * payload, uint16_t payload_length);
 
 // Set the callback when a specific yasp command is received
-void register_yasp_command(command_callback callback, uint8_t command);
+void register_yasp_command(command_callback callback, uint16_t command);
 
 // Attach yasp to the serial port
 void yasp_init();
 
-void send_yasp_command(uint8_t cmd, uint8_t * payload, uint16_t payload_length, uint8_t ack);
+void send_yasp_command(uint16_t cmd, uint8_t * payload, uint16_t payload_length, uint8_t ack);
 
 #endif // _YASP_H_
